Setting awc_mixing_factor as a float

+++ floris/core/wake_velocity/empirical_gauss.py
@@ -297,7 +297,7 @@ def awc_added_wake_mixing(
     awc_mode_i = awc_mode_i[:,:,0,0]
 
     # TODO: Add TI in the mix, finetune amplitude/freq effect
-    awc_mixing_factor = np.zeros_like(awc_amplitude_i)
+    awc_mixing_factor = np.zeros_like(awc_amplitude_i,dtype=float)
     helix_mask = awc_mode_i == 'helix'
 
     awc_mixing_factor[helix_mask] = (
